Server: Move JWToken service to a more descriptive folder and create interface to expose services for JWTService.
